refactor: 命名の変更とそれに伴う修正
