--- tests/TestConverter.py
@@ -1,6 +1,0 @@
-from converter import Converter
-
-
-class TestConverter(object):
-    def test_output_pdf(self):
-        Converter([]).to_pdf()
