+++ tests/test_pdf_converter.py
@@ -1,0 +1,6 @@
+from infra.pdf_converter import PdfConverter
+
+
+class TestConverter(object):
+    def test_output_pdf(self):
+        PdfConverter([]).execute()
